+++ packages/examples-hooks/src/01-dustbin/single-target-in-iframe/index.tsx
@@ -6,8 +6,8 @@ import {
 	DragDropContextProvider,
 } from 'react-dnd'
 import HTML5Backend from 'react-dnd-html5-backend'
-import Dustbin from '../Single Target/Dustbin'
-import Box from '../Single Target/Box'
+import Dustbin from '../single-target/Dustbin'
+import Box from '../single-target/Box'
 import { isDebugMode } from '../../isDebugMode'
 
 const {
