+++ packages/examples/src/03-nesting/drag-sources/SourceBox.tsx
@@ -7,7 +7,6 @@ import {
 	DragSourceConnector,
 } from 'react-dnd'
 import Colors from './Colors'
-import { SourceBox } from '../../01 Dustbin/Stress Test'
 
 const style: React.CSSProperties = {
 	border: '1px dashed gray',
